Handle unqualified imports in custom Triton kernels
Summary: If in a custom (user-written) Triton kernel an externally imported symbol is used directly, we need to codegen the corresponding import outside the kernel body in the Python wrapper. E.g., if the user code has this:

```
    from triton.language.extra.cuda.libdevice import fast_dividef

    triton.jit
    def my_kernel(...):
        ...
        x = fast_dividef(...)
        ...
```

The `from triton.language.extra.cuda.libdevice import fast_dividef` line needs to be carried over together with the `my_kernel` function. The PR adds this.

Test Plan:

```
$ python test/inductor/test_triton_kernels.py
...
----------------------------------------------------------------------
Ran 464 tests in 113.512s

OK
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: ac65d9a6ed5b6bb9c364e74acfe77db4e82a5804
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123703