Add ability to save TORCH_COMPILE_DEBUG logs for CI failures
Summary: The idea is that we can enroll certain benchmarks to a) enable TORCH_COMPILE_DEBUG=1, and b) save those logs in test/debug in case of a failure. We can then update action.yml to upload test/debug/ to S3 whenever it exists.

Test Plan:
See artifacts generated when I force a benchmark failure:
https://hud.pytorch.org/pr/124234 (e.g., debug-test-inductor_torchbench-2-2-linux.g5.4xlarge.nvidia.gpu_23904773386.zip)

ghstack-source-id: 6fe640e1b46aefe1554d7da6068acd955fec27dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124234