[custom_op] register_autograd raises NYI on kwarg-only args
It's unclear how to actually handle kwarg-only args, so we raise NYI for
now and hope that there aren't too many use cases for this.

Also, an autograd impl (that raises error on backward) gets unconditionally
registered for the op. We fix that to handle kwarg-only args (assuming the user
did not call register_autograd) by flattening the kwarg-only args and
sending them through the autograd.Function.

Test Plan:
- new tests

ghstack-source-id: f30e7d19b49f4e6e8d1dd3f638801a945ace84ee
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124638