Add slice_strict
slice_strict is equivalent to slice, except that it does not accept
start/end that are out of bounds.  This makes it more friendly for
unbacked SymInts.

To avoid blast radius, slice_strict is only called by narrow when
in PT2.  This may still have some blast radius on export which will
need to be resolved.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: f0a62548ccdaaf2f4caa3ead66c9a03c1293f418
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124683