Verify types in custom op schemas
Before this PR, we didn't check that types in a schema were valid. This
is because TorchScript treats unknown types as type variables.

This PR checks types in a schema for the TORCH_LIBRARY APIs. To do this,
we add an `allow_typevars` flag to parseSchema so that TorchScript can
use allow_typevars=True. We also add some error messages for common
mistakes (e.g. using int64_t or double in schema).

Test Plan:
- new tests

ghstack-source-id: 8e45ebe0b9c6af8b79b183c0b14d4f93762e2045
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124520