Make c10d_functional ops call into _c10d_functional ops
This PR removes the legacy impls of c10d_functional ops which are now irrelevant. For backward compatibility purpose, c10d_functional ops now call into _c10d_functional ops.

We also changed c10d_functional ops to be CompositeExplicitAutograd, so that when traced, only _c10d_functional ops appear in the graph. After this, we'll be able to remove the Inductor IR for the legacy functional collectives.

ghstack-source-id: fbf2cb62519d5484167a24ee23eb7e480d568e86
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124979