Add --warm-start-latency to benchmark harness
Summary: This change introduces a new flagg to perform a "warm start" test from the benchmark harness. The idea is to test a model twice: first with a fresh inductor cache (i.e., a "cold start"), and then a second run in a fresh process with the cache available (i.e. a "warm start"). We can later add this mode to CI runs to collect compile times for warm start.

ghstack-source-id: b57df24736231a6e5eaa0728bd06e6fa90de2e04
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125353