Add FlopCounter implementations for NestedTensor SDPA kernels
This adds implementations for:
* _flash_attention_forward
* _efficient_attention_forward
* _flash_attention_backward
* _efficient_attention_backward

The implementations: (a) effectively "pad" the nested tensors to max seq len, and (b) call the normal flop calculation for SDPA assuming that we had the max seq len. So, this is actually an upper bound for the FLOPs, assuming that the nested tensor had all sequences at max seq len.

ghstack-source-id: 764b1057331e7fa42b067fb7d91d576514f34377
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125712