[benchmarking] Suppress csv creation on cold-start phase of --warm-start-latency
Summary: It seems that most (all?) of our utilities for examining benchmark output expect single-line entries per benchmark. The way the --warm-start-latency flag is currently implemented, it means that we'll see two entries for every benchmark run (one for the warm-up run and one for the actual run). This PR adds a --disable-output flag that we can use for the first run to suppress populating the csv. This way, the existing utilities like `benchmarks/dynamo/check_accuracy.py` will function without any changes.

ghstack-source-id: 45248307604993c7586b3a23c11ff4882ac44a71
Pull Request resolved: https://github.com/pytorch/pytorch/pull/125953